adjust_path Remove non existent directories from path.
This is a workaround for a WSL2 related problem.
If there is a windows directory symlink into a WSL2 UNC mount
e.g. ubuntu_user_on_wsl -> \\wsl.localhost\Ubuntu\home\ubuntu_user
then none of the Ubuntu (and presumably other) OS's hosted
will be able to use the windows directory symlink.

In Python3 this manifests as an Input/Output error (error 5) which
breaks any tool trying to access that file.
The simplest fix is not to have unusable directories on the PATH
in the first place.